# docs/api/openapi.yaml
swagger: '2.0'
info:
  title: netobs unified API
  version: 2.5.6
paths:
...
  /api/v1/nodes:
    get:
      description: 노드별 이름, uid, 내부/외부 IP, Ready 상태, 역할(control-plane 등), 버전, capacity(cpu/memory/gpu)를
        kube-state-metrics 기반으로 돌려준다. 다른 API의 node 라벨과 동일 키로 매핑한다.
      produces:
      - application/json
      tags:
      - inventory
      summary: 노드 인벤토리
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodesResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/pods:
    get:
      description: '파드별 namespace, 이름, uid, pod IP, host IP, node, workload(created_by),
        priority, phase, qos, 관측 커버리지(observed)를 kube-state-metrics 기반으로 돌려준다. observed
        는 netobs 의 eBPF 시리즈가 존재하는지로, false 면 unobserved_reason 이 미관측 사유를 분류한다 (#320,
        #342): agent_absent 는 노드에 netobs 미배치, host_network 는 pod IP 가 host IP 와 같아
        IP 귀속이 성립하지 않는 상태로 cgroup 힌트가 학습되면 live 로 전환될 수 있는 시점 의존 분류다. no_traffic 은
        agent 의 소켓 스캔이 pod netns 의 무소켓을 증명한 상태 (unix socket 전용 pod 등) 로 시리즈 부재가 정상이고,
        no_data 는 관측 가능한데 시리즈 부재다. 종료 pod (phase Succeeded/Failed) 는 telemetry 부재가
        정상이라 observed=false 에 사유가 생략된다 (#314). ?namespace 로 필터한다. 다른 API의 src_namespace/src_pod/pod_uid와
        동일 키로 매핑한다.'
      produces:
      - application/json
      tags:
      - inventory
      summary: 파드 인벤토리
      parameters:
      - type: string
        description: namespace 필터 (생략 시 전체)
        name: namespace
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.PodsResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
definitions:
...
  internal_correlation_api.NodeCapacity:
    type: object
    properties:
      cpu:
        type: number
      gpu:
        type: number
      memory_bytes:
        type: number
...
  internal_correlation_api.NodeInventory:
    type: object
    properties:
      capacity:
        $ref: '#/definitions/internal_correlation_api.NodeCapacity'
      external_ip:
        type: string
      internal_ip:
        type: string
      kernel_version:
        type: string
      kubelet_version:
        type: string
      name:
        type: string
      os_image:
        type: string
      ready:
        type: boolean
      roles:
        description: 'Roles 는 kube_node_role 기반 노드 역할 (control-plane 등) 이다 (#248).
          역할 라벨이 없는

          worker 노드는 생략된다.'
        type: array
        items:
          type: string
      runtime_version:
        type: string
      uid:
        type: string
...
  internal_correlation_api.NodesResponse:
    type: object
    properties:
      generated_at:
        type: string
      nodes:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.NodeInventory'
...
  internal_correlation_api.PodInventory:
    type: object
    properties:
      host_ip:
        type: string
      namespace:
        type: string
      node:
        type: string
      observed:
        description: 'Observed 는 eBPF 관측 커버리지다 (#248). netobs 가 상시 수집하는 netobs_pod_bytes_total

          시리즈가 이 pod 에 존재하면 true, 없으면 no-data 로 판정한다.'
        type: boolean
      phase:
        type: string
      pod:
        type: string
      pod_ip:
        type: string
      priority_class:
        type: string
      qos_class:
        type: string
      uid:
        type: string
      unobserved_reason:
        description: 'UnobservedReason 은 미관측 사유 분류다 (#320, #342). agent_absent (노드에
          netobs 미배치),

          host_network (pod IP 가 host IP 와 같아 IP 귀속이 성립하지 않음), no_traffic (netns 소켓

          테이블이 비어 네트워크 미사용이 증명됨, unix socket 전용 pod 등), no_data (관측 가능한데

          시리즈 부재) 로 구분한다. host_network 는 확정이 아닌 시점 의존 분류로, cgroup 힌트 (#228)

          가 TCP 트래픽에서 학습되면 observed=true (live) 로 전환될 수 있다. observed=true 와

          종료/Unknown phase 는 생략된다.'
        type: string
      workload_kind:
        type: string
      workload_name:
        type: string
...
  internal_correlation_api.PodsResponse:
    type: object
    properties:
      generated_at:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.PodInventory'
...
  netobs_internal_apicommon.ErrorBody:
    type: object
    properties:
      error:
        $ref: '#/definitions/netobs_internal_apicommon.ErrorDetail'
  netobs_internal_apicommon.ErrorDetail:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...